Move component to top level
