+++ src/app/feed/feed.page.ts
@@ -16,7 +16,7 @@ import {
 } from '@ionic/angular/standalone';
 import { addIcons } from 'ionicons';
 import { bookmark, bookmarkOutline, chevronForward, chevronUpOutline, ellipsisVertical, filterOutline, shareSocialOutline } from 'ionicons/icons';
-import { ArticleListComponent } from '../lib/article-list/article-list.component';
+import { ArticleListComponent } from '../article-list/article-list.component';
 import { formatDateAsDay, formatDateAsLong, formatDateRelative } from '../lib/date-utils';
 import { BookmarkService } from '../services/bookmark.service';
 import { FeedService } from '../services/feed.service';
